text from / to clipboard(pasteboard)+ bug fixes

+++ include/util/segList.h
@@ -111,7 +111,7 @@ friend class segList;
   void **freeSpc;   /// list with all free spaces in segment ( unitSize* segSize total space)
   int freeSpcPeak;  /// peak of freeSpc. acts like some kind of stack: u pick a free space from the top/ u put the space taken back
 
-  unsigned int timeIdle;    /// if 0, segment is in use. Else it holds the time @ start idling (used internally by segList::checkIdle())
+  uint64_t timeIdle;    /// if 0, segment is in use. Else it holds the time @ start idling (used internally by segList::checkIdle())
 
   _segment();
   ~_segment();
